refactor: 멀티모듈 구조로 리펙토링
