+++ backend/admin-web/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    implementation project(':domain')
+    
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.data:spring-data-commons'
+    
+    // Swagger/OpenAPI
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
+    
+    // Test
+    testImplementation 'org.springframework.security:spring-security-test'
+}
