+++ backend/admin-web/src/main/java/com/techbloghub/admin/dto/AdminBatchDeleteRequest.java
