+++ backend/domain/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+    // Core domain should have minimal dependencies
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'com.fasterxml.jackson.core:jackson-databind'
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
+    
+    // Minimal Spring dependencies for domain services
+    implementation 'org.springframework:spring-context'
+    implementation 'org.springframework:spring-tx'
+    implementation 'org.springframework.data:spring-data-commons'
+}
