+++ backend/domain/src/main/java/com/techbloghub/domain/model/SearchCondition.java
