+++ backend/persistence/build.gradle
@@ -1,0 +1,23 @@
+dependencies {
+    implementation project(':domain')
+    
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+    
+    // RSS Feed Parser
+    implementation 'com.rometools:rome:2.1.0'
+    implementation 'com.rometools:rome-utils:2.1.0'
+    
+    // Querydsl
+    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+    annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
+    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+    
+    // Database
+    runtimeOnly 'org.postgresql:postgresql'
+    
+    // Test
+    testImplementation 'org.testcontainers:postgresql'
+    testImplementation 'org.testcontainers:junit-jupiter'
+}
