+++ backend/settings.gradle
@@ -1,1 +1,7 @@
 rootProject.name = 'techbloghub-backend'
+
+include 'domain'
+include 'api-web'
+include 'admin-web'
+include 'persistence'
+include 'bootstrap'
